+++ frontend/src/Containers/Main/Main.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import classNames from "classnames";
+import Header from "../../Components/Header/Header";
+import css from "./Main.module.css";
+
+const Main = () => {
+  return (
+    <div className={css.main}>
+      <Header />
+      <div className={css.main__wrapper}>
+
+      </div>
+    </div>
+  );
+};
+
+export default Main;
